Move all event bus based logic into event bus class and simplify up event base
Remove legacy event metaprogrammed class
