+++ lib/cucumber/core/event/test_case_finished.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../event'
+require_relative 'base'
 
 module Cucumber
   module Core
